+++ SoftLayer/fixtures/SoftLayer_Location.py
@@ -9,3 +9,4 @@ getDataCenters = [
         "longName": "San Jose 1",
         "name": "sjc01"
     }]
+getDatacenters = [{'id': 1854895, 'name': 'dal13', 'regions': [{'keyname': 'DALLAS13'}]}]