+++ SoftLayer/fixtures/SoftLayer_Network_Vlan.py
@@ -1,1 +1,9 @@
-getObject = {'primaryRouter': {'datacenter': {'id': 1234}}}
+getObject = {
+    'primaryRouter': {
+        'datacenter': {'id': 1234, 'longName': 'TestDC'},
+        'fullyQualifiedDomainName': 'fcr01.TestDC'
+    },
+    'id': 1234,
+    'vlanNumber': 4444,
+    'firewallInterfaces': None
+}